feat: 优化角色权限缓存逻辑，调整用户认证信息结构体字段类型，更新测试用例以反映更改

+++ utils/claims.go
@@ -13,8 +13,8 @@ import (
 // 自定义声明结构体
 type CustomClaims struct {
 	UserID      int64    `json:"userId"`
-	DeptID      int64    `json:"deptId"`
-	DataScope   int      `json:"dataScope"`
+	DeptID      int      `json:"deptId"`
+	DataScope   int8     `json:"dataScope"`
 	Authorities []string `json:"authorities"`
 	jwt.RegisteredClaims
 }
